requirements.txt
torch
torchaudio
torchvision
Pillow
trimesh
meshio
matplotlib
numpy
scipy
polyscope
sounddevice